backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6
beautifulsoup4==4.12.2
selenium==4.11.2
webdriver-manager==4.0.0
google-generativeai==0.3.1
pydantic==2.3.0
python-dotenv==1.0.0